Add `exit` event to terminate execution
Also, the events code was moved to its own file.
