+++ modules/edu.gemini.seqexec.engine/src/main/scala/edu/gemini/seqexec/engine/Engine.scala
@@ -1,47 +1,14 @@
 package edu.gemini.seqexec.engine
 
+import Event._
+import Sequence._
 import scalaz._
-import Scalaz._
+import scalaz.Scalaz._
 import scalaz.concurrent.Task
 import scalaz.stream.Process
 import scalaz.stream.async.mutable.Queue
-import edu.gemini.seqexec.engine.Sequence._
 
 object Engine {
-  /**
-    * Anything that can go through the Event Queue.
-    *
-    */
-  sealed trait Event
-  case class EventUser(ue: UserEvent) extends Event
-  case class EventSystem(se: SystemEvent) extends Event
-
-  sealed trait UserEvent
-  case object Start extends UserEvent
-  case object Pause extends UserEvent
-  case class AddStep(a: Step) extends UserEvent
-
-  val start: Event = EventUser(Start)
-  val pause: Event = EventUser(Pause)
-  def addStep(ste: Step): Event = EventUser(AddStep(ste))
-
-  sealed trait SystemEvent
-  // when an action is completed even if it belongs to a set of
-  // parallel actions.
-  case object Completed extends SystemEvent
-  // when an action failed
-  case object Failed extends SystemEvent
-  // when a set of parallel actions is completed.
-  case object Synced extends SystemEvent
-  case object SyncFailed extends SystemEvent
-  case object Finished extends SystemEvent
-
-  val completed: Event = EventSystem(Completed)
-  val failed: Event = EventSystem(Failed)
-  val synced: Event = EventSystem(Synced)
-  val syncFailed: Event = EventSystem(SyncFailed)
-  val finished: Event = EventSystem(Finished)
-
   /**
     * Input Sequence and Status clubbed together
     *
@@ -180,4 +147,9 @@ object Engine {
   def add(ste: Step): Telescope[SeqStatus] =
     MonadState[Telescope, SeqStatus].modify(_.leftMap(ste :: _)) *>
       MonadState[Telescope, SeqStatus].get
+
+  def close(queue: Queue[Event]): Telescope[SeqStatus] = {
+    queue.close.liftM[TelescopeStateT] *>
+      MonadState[Telescope, SeqStatus].get
+  }
 }
